requirements.txt
confluent-kafka==1.6.0
avro-python3==1.10.1
requests==2.25.1
fastavro==1.3.2